[IMP] coupon_limit, sale_coupon_limit split
Allow coupon_limit settings to be used by other loyalty modules
