+++ coupon_limit/models/__init__.py
@@ -1,3 +1,1 @@
-from . import coupon_coupon
-from . import coupon_program
 from . import coupon_rule
