+++ coupon_limit/readme/DESCRIPTION.rst
@@ -1,3 +1,2 @@
-This module allows to configure a limit on the times a promotion can be applied. Two
-limits can be configured: customer and salesman. Those limits apply to either programs
-or coupons.
+This module allows to configure a limit on the times a promotion can be applied. It's a
+technical base to be extended with top level functionality (sales, pos).
