+++ sale_coupon_limit/README.rst
@@ -1,0 +1,122 @@
+============
+Coupon Limit
+============
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Production%2FStable-green.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Production/Stable
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fsale--promotion-lightgray.png?logo=github
+    :target: https://github.com/OCA/sale-promotion/tree/15.0/sale_coupon_limit
+    :alt: OCA/sale-promotion
+.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
+    :target: https://translation.odoo-community.org/projects/sale-promotion-15-0/sale-promotion-15-0-sale_coupon_limit
+    :alt: Translate me on Weblate
+.. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
+    :target: https://runbot.odoo-community.org/runbot/296/15.0
+    :alt: Try me on Runbot
+
+|badge1| |badge2| |badge3| |badge4| |badge5| 
+
+This module allows to configure a limit on the times a promotion can be applied. Two
+limits can be configured: customer and salesman. Those limits apply to either programs
+or coupons.
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Configuration
+=============
+
+To configure customer limits:
+
+#. Go to *Sales > Catalog > Coupon Programs* and select or create a new one.
+#. Set the *Maximum Customer Applications* to the number of times a coupon can be used
+   by a customer.
+
+NOTE: The customer limit is applied at commercial entity level, not for each contact
+inside the entity.
+
+To configure salesmen limits:
+
+#. Go to *Sales > Catalog > Coupon Programs* and select or create a new one.
+#. Add salesmen to the *Salesmen Limits* list and their maximum number of applications.
+#. You can add different limits to different salesmen groups.
+#. If you want to constrain the use of the promotion to the salesmen list, set the
+   option *Strict limit* on, so any other salesman won't be able to apply the promotion.
+
+Usage
+=====
+
+Once the program limits are configured, apply the programs as usual in your sale orders.
+
+Once the limit for a customer or a salesman is reached, if we try to apply a promotion:
+
+- A code promotion will raise an error.
+- A program with no code won't be applied.
+- A coupon belonging to a limited program will raise an error.
+- A promotion applied on the next order won't generate the coupon.
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/sale-promotion/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/OCA/sale-promotion/issues/new?body=module:%20sale_coupon_limit%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* Tecnativa
+
+Contributors
+~~~~~~~~~~~~
+
+* `Tecnativa <https://www.tecnativa.com>`_:
+
+  * Pedro M. Baeza
+  * David Vidal
+
+* `Akretion <https://www.akretion.com>`_:
+
+  * Florian Mounier
+
+Maintainers
+~~~~~~~~~~~
+
+This module is maintained by the OCA.
+
+.. image:: https://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: https://odoo-community.org
+
+OCA, or the Odoo Community Association, is a nonprofit organization whose
+mission is to support the collaborative development of Odoo features and
+promote its widespread use.
+
+.. |maintainer-chienandalu| image:: https://github.com/chienandalu.png?size=40px
+    :target: https://github.com/chienandalu
+    :alt: chienandalu
+
+Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
+
+|maintainer-chienandalu| 
+
+This module is part of the `OCA/sale-promotion <https://github.com/OCA/sale-promotion/tree/15.0/sale_coupon_limit>`_ project on GitHub.
+
+You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
