+++ sale_coupon_limit/models/__init__.py
@@ -1,0 +1,3 @@
+from . import coupon_coupon
+from . import coupon_program
+from . import coupon_rule
