+++ sale_coupon_limit/readme/CONFIGURE.rst
