+++ sale_coupon_limit/readme/CONTRIBUTORS.rst
@@ -1,0 +1,8 @@
+* `Tecnativa <https://www.tecnativa.com>`_:
+
+  * Pedro M. Baeza
+  * David Vidal
+
+* `Akretion <https://www.akretion.com>`_:
+
+  * Florian Mounier
