+++ setup/sale_coupon_limit/odoo/addons/sale_coupon_limit
@@ -1,0 +1,1 @@
+../../../../sale_coupon_limit
